+++ src/components/com-nav-submenu/com-nav-submenu.tsx
@@ -10,8 +10,12 @@ import {
     Event,
     EventEmitter, Method
 } from '@stencil/core';
-import { MenuOpenedEventDetail } from '../com-nav/com-nav-interface';
+import {MenuOpenedEventDetail} from '../com-nav/com-nav-interface';
+import {isEnterKey, isEscapeKey, isSpaceKey} from "../../utils/helpers";
+import {registerCloseHandlerForFocusOutEventOn} from "../com-nav/com-nav-utils";
 
+
+const WCS_COM_NAV_CATEGORY = 'WCS-COM-NAV-CATEGORY';
 
 @Component({
     tag: 'wcs-com-nav-submenu',
@@ -32,6 +36,11 @@ export class ComNavSubmenu implements ComponentInterface {
      */
     @Event() wcsClickOnFinalAction: EventEmitter<void>;
 
+    componentWillLoad(): Promise<void> | void {
+        const slottedCategoryItems = this.el.querySelectorAll(':scope > wcs-com-nav-category:not([slot])');
+        registerCloseHandlerForFocusOutEventOn<HTMLWcsComNavCategoryElement>(slottedCategoryItems, WCS_COM_NAV_CATEGORY);
+    }
+
     /**
      * If the user clicks outside the menu, we close it
      */
@@ -44,6 +53,17 @@ export class ComNavSubmenu implements ComponentInterface {
     onSubmenuOpened(event: CustomEvent<MenuOpenedEventDetail>) {
         // If the clicked menu is not this component, we close it
         if (event.detail.menuElement !== this.el) {
+            this.menuOpen = false;
+        }
+    }
+
+    /**
+     * Close the menu when escape is pressed
+     * @param _event keydown event
+     */
+    @Listen('keydown', {target: 'window'})
+    onEscapeKeyDown(_event: KeyboardEvent) {
+        if (isEscapeKey(_event) && this.menuOpen) {
             this.menuOpen = false;
         }
     }
@@ -73,8 +93,29 @@ export class ComNavSubmenu implements ComponentInterface {
         this.wcsSubmenuOpened.emit({menuElement: this.el})
     }
 
+    /**
+     * Handle key down on menu items
+     * @param _event the keyboard event
+     * @private
+     */
+    private handleMenuItemsKeyDown(_event: KeyboardEvent) {
+        if ((isSpaceKey(_event)) || isEnterKey(_event)) {
+            this.handleMenuItemsClick(_event);
+        }
+    }
 
-    private handleMenuItemsClick(evt: MouseEvent) {
+    /**
+     * Open the menu if it is closed and closed the menu if it is already opened
+     * @param _event the keyboard event
+     * @private
+     */
+    private handleMenuKeyDown(_event: KeyboardEvent) {
+        if ((isSpaceKey(_event)) || isEnterKey(_event)) {
+            this.menuOpen = !this.menuOpen;
+        }
+    }
+
+    private handleMenuItemsClick(evt: UIEvent) {
         if ((evt.target as HTMLElement).tagName === 'A') {
             this.close();
             this.wcsClickOnFinalAction.emit();
@@ -88,7 +129,7 @@ export class ComNavSubmenu implements ComponentInterface {
      */
     @Listen('wcsCategoryItemClicked')
     // @ts-ignore
-    private wcsCategoryItemClickedHandler(_: CustomEvent<MouseEvent>) {
+    private wcsCategoryItemClickedHandler(_: CustomEvent<UIEvent>) {
         // If a category item is clicked, we close the submenu drawer;
         this.close();
     }
@@ -97,7 +138,10 @@ export class ComNavSubmenu implements ComponentInterface {
     render(): any {
         return (
             <Host onClick={evt => this.onClick(evt)}>
-                <div onClick={_ => this.menuOpen = !this.menuOpen} class="menu-button">
+                <div tabindex={screen.width < 576 ? "-1" : "0"}
+                     onClick={_ => this.menuOpen = !this.menuOpen}
+                     onKeyDown={evt => this.handleMenuKeyDown(evt)}
+                     class="menu-button">
                     <span class="label">{this.label}</span><span class="arrow-container"><span
                     class="arrow-icon" data-open={this.menuOpen}>&#xf107;</span></span>
                 </div>
@@ -108,7 +152,9 @@ export class ComNavSubmenu implements ComponentInterface {
                                 <h3>{this.panelTitle}</h3>
                                 <p>{this.panelDescription}</p>
                             </div>
-                            <div class="menu-items" onClick={(evt) => this.handleMenuItemsClick(evt)}>
+                            <div class="menu-items"
+                                 onClick={(evt) => this.handleMenuItemsClick(evt)}
+                                 onKeyDown={evt => this.handleMenuItemsKeyDown(evt)}>
                                 <slot/>
                             </div>
                         </div>